make repository pattern and make controller auth

+++ database/migrations/2022_06_30_124722_create_orders_table.php
@@ -17,8 +17,8 @@ class CreateOrdersTable extends Migration
             $table->unsignedBigInteger('currency_id')->comment('شناسه ارز');
             $table->foreign('currency_id')->references('id')->on('currencies');
 
-            $table->bigInteger('amount')->comment('میزان دارایی');
-            $table->tinyInteger('fee')->default(0)->comment('فی');
+            $table->decimal('amount', 16, 8)->comment('میزان دارایی');
+            $table->decimal('fee', 16, 8)->default(0)->comment('فی');
             $table->timestamps();
             $table->softDeletes();
         });
